+++ justify-tests/src/test/java/org/leadpony/justify/tests/extra/json/Draft04OfficialTest.java
@@ -71,6 +71,7 @@ public class Draft04OfficialTest extends AbstractOfficialTest {
             "optional/ecmascript-regex.json",
             "optional/format.json",
             "optional/zeroTerminatedFloats.json",
+            "optional/non-bmp-regex.json",
     };
 
     public static Stream<TestCase> optional() {