+++ CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 3.0.0-RC2 - 2020-05-12
 ### Added
-- new API `JsonValidationService.newInstance(JsonProvider)` 
+- A new API `JsonValidationService.newInstance(JsonProvider)`.
 
 ## Fixed
-- A bug that was causing `uniqueItems` having `false` to produce wrong validation result.
+- A bug that was causing `uniqueItems` keyword with `false` value to produce wrong validation result.
 
 ## 3.0.0-RC1 - 2020-04-27
 ### Changed